Force zIndex for multiple menu inside eachother
I have a navigation where I needed a menu setup with options when you right click anywhere on it. And when you right click on the buttons in the navigation, we needed a different set of options.

I know there is the ability to change the options with the current plugin, but the quick and easy way for us to provide what we needed without having to 'check' and 'swap' between menus, was to create multiple instances, one on the main navigation, and one on the specific buttons.

With that, there's an error where the parent menu seems to have the higher zIndex, instead of the buttons. I went ahead and made a simple adjustment to your code to provide the ability to handle this type of stacking instance with your context menus.